+++ src/pdf.js
@@ -36,7 +36,6 @@ import {
   makeArr,
   makeMap,
   makeObj,
-  MathClamp,
   normalizeUnicode,
   OPS,
   PasswordResponses,
@@ -84,6 +83,7 @@ import { DrawLayer } from "./display/draw_layer.js";
 import { GlobalWorkerOptions } from "./display/worker_options.js";
 import { HighlightOutliner } from "./display/editor/drawers/highlight.js";
 import { isValidExplicitDest } from "./display/api_utils.js";
+import { MathClamp } from "./shared/math_clamp.js";
 import { SignatureExtractor } from "./display/editor/drawers/signaturedraw.js";
 import { TextLayer } from "./display/text_layer.js";
 import { TextLayerImages } from "./display/text_layer_images.js";
